Harden the pipeline CLIs: skip bad images, guard missing DBs, portable BLOBs
- ingest.py skips a missing/corrupt image with a SKIP line instead of
  aborting the whole batch with a traceback after the model has loaded,
  and rejects --caption with multiple paths (it would silently give
  every image the same caption and text embedding)
- search.py and export_web.py check the database file exists before
  connecting, so a typo'd --db no longer creates a stray empty sqlite
  file; the check also runs before the 600 MB model load
- search.py validates -k >= 1 (negative k silently returned a wrong
  result set via Python slicing) and errors early on a missing --image
- export_web.py gains the same --db flag as ingest/search
- db.py spells the BLOB dtype as explicit little-endian '<f4' so a
  gallery written on one machine decodes identically everywhere
- features.py normpaths stored paths so 'images/cat.jpg' and
  './images/cat.jpg' hit the same UNIQUE row instead of duplicating
- ClipEmbedder imports are deferred to first real use, so score(),
  FeatureExtractor with an injected encoder, and all CLI error paths
  run without torch/transformers installed
- all four CLIs use RawDescriptionHelpFormatter so --help keeps the
  pipeline diagrams readable

+++ db.py
@@ -36,11 +36,13 @@ def connect(path: str = DB_PATH) -> sqlite3.Connection:
 
 
 def to_blob(vec) -> bytes:
-    return np.asarray(vec, dtype=np.float32).tobytes()
+    # "<f4" = little-endian float32, spelled out so a gallery.sqlite written
+    # on one machine decodes to the same floats on any other
+    return np.asarray(vec, dtype="<f4").tobytes()
 
 
 def from_blob(blob: bytes) -> np.ndarray:
-    return np.frombuffer(blob, dtype=np.float32)
+    return np.frombuffer(blob, dtype="<f4")  # read-only view over the bytes
 
 
 def add_image(con, path, caption, tags, image_emb, text_emb, fused_emb):
